word wrap logic for item descriptions
#3

to cover the changes briefly: when the item icons are drawn the height
to draw them was previously hardcoded. implemented a new draw text
function that will optionally do word wrap and returns the height of the
text drawn. keeps a "text height" property for when item icons are
drawn. additionally, added this as an option in option_picker.py

i also tacked on current floor as a property that gets parsed from the
log.