Removed old code and changed guppy tracking
I removed several references to the various lists that were kept to
filter out stuff.  I also swapped out the guppy list with a guppy count
since we can always loop over the collected_item list to determine if
we're *really* guppy or not.  I did not due this change in a dev branch
because it's a bunch of minor changes.