Implement __hash__ for Item
Fixes #63.
Python uses __hash__ method from objects to check if it's in a set.
For Item we can just hash the id since two items are equal if their ids are equal.

For this issue, Guppy's paw was added twice to the set, therefore displaying a
guppy counter of 2 instead of 1.

Modify the caching system so that last floor is updated on a previously cached item.