Create (De)serialization for trackers' object.
Objects to be serialized must derive from "Serializable", and define a class member 'serialize',
containing a list of members and theirs types to be (de)serialize.
E.g. [('floor_id', str), ('curse', int)] for a floor.

Serialization is handled entirely by the base class, deserialization occurs in two parts :
    - Members and type checking in the base class
    - Data validation and Object creation in the deribed class
If deserialization fails at some point, a message is written to the log and 'None' is returned.

Serialization on the whole state can be achieved via a simple call to json.dumps :
    `json.dumps(state, cls=TrackerStateEncoder, sort_keys=True)`
Deserialization can be achieved by giving a json dictionary to the TrackerState :
    `state = TrackerState.from_json(json_dic)`