Make overlapping items stretch to the full width.
If there is enough room for all items on a single row we want them "floating left".
If we start making items overlap with each others, we want them to stretch to the full width of the window.

The second point was buggy, I rewrote the reflow function to compute beforehand the number of items per row while minimising the overlap.
It eventually remains unused pixels, which are evenly distributed per item to stretch to the full width of the window.