change run reset behavior to fix "continue" bug
in 1.2.0 I changed it to trigger resets when it sees a log message about the seed.
that message is included when you continue a run, so that's no good.
the reason I did it is that antibirth doesn't have the "start room" message
I made antibirth keep using the seed message, and other versions use "floor = 1" message
which kilburn told me he will add to the next version of antibirth